test(LottoTest): Lotto 관련 단위테스트 하나로 통합
