+++ src/test/java/lotto/domain/LottoTest.java
@@ -9,16 +9,30 @@ import java.util.List;
 import static lotto.LottoConstants.LOTTO_SIZE;
 import static lotto.message.ErrorMessage.LOTTO_NUMBERS_DUPLICATE;
 import static lotto.message.ErrorMessage.LOTTO_SIZE_ERROR;
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatCode;
+import static org.assertj.core.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 public class LottoTest {
 
-    @DisplayName("로또는 6개의 숫자로 이루어져야 한다.")
+    @DisplayName("로또 번호의 개수가 6개가 넘어가면 예외가 발생한다.")
+    @Test
+    void 로또_번호의_개수가_6개가_넘어가면_예외가_발생한다() {
+        assertThatThrownBy(() -> new Lotto(List.of(1, 2, 3, 4, 5, 6, 7)))
+                .isInstanceOf(IllegalArgumentException.class);
+    }
+
+    @DisplayName("로또 번호에 중복된 숫자가 있으면 예외가 발생한다.")
+    @Test
+    void 로또_번호에_중복된_숫자가_있으면_예외가_발생한다() {
+        assertThatThrownBy(() -> new Lotto(List.of(1, 2, 3, 4, 5, 5)))
+                .isInstanceOf(IllegalArgumentException.class);
+    }
+
+    @DisplayName("로또의 숫자가 6개 미만이면 예외가 발생한다.")
     @Test
     void lottoSizeTest() {
         // given
-        List<Integer> numbers = List.of(12, 5, 7, 8, 1, 9, 27);
+        List<Integer> numbers = List.of(12, 5, 7, 8, 1);
 
         // when & then
         assertThatCode(() -> new Lotto(numbers))
@@ -27,7 +41,7 @@ public class LottoTest {
 
     }
 
-    @DisplayName("로또에 저장되는 6개의 숫자는 중복되면 안된다.")
+    @DisplayName("로또에 저장되는 6개의 숫자가 중복되면 예외가 발생한다")
     @Test
     void duplicateTest() {
         // given
